requirements.txt
python-telegram-bot==13.7
redis==3.5.3
python-dotenv==0.19.2
